refactor(edusharing-asset): move logic to custom hook
